+++ src/Raven.CodeAnalysis/Syntax/InternalSyntax/Parser/Parsers/TypeDeclarationParser.cs
@@ -202,7 +202,6 @@ internal class TypeDeclarationParser : SyntaxParser
             SyntaxKind.StructKeyword or
             SyntaxKind.InterfaceKeyword or
             SyntaxKind.ExtensionKeyword or
-            SyntaxKind.TraitKeyword or
             SyntaxKind.EnumKeyword or
             SyntaxKind.UnionKeyword or
             SyntaxKind.DelegateKeyword or