+++ test/Raven.CodeAnalysis.Tests/Syntax/Parser/ParserTokenInsertionFuzzTests.cs
@@ -31,7 +31,6 @@ public class ParserTokenInsertionFuzzTests
         ")",
         "[",
         "]",
-        "trait",
         "class",
         "namespace",
         "import",